+++ doc/whats_new.rst
@@ -12,6 +12,7 @@ New
 * Use :func:`nistats.reporting.make_glm_report` to easily generate HTML reports from fitted first and second level models and contrasts.
 * New dataset fetcher, :func:`nistats.datasets.fetch_language_localizer_demo_dataset` , BIDS 1.2 compatible.
 * New example showcasing the use of a GLM to get beta maps for decoding experiments (aka beta-regression).
+* Addition of all-resolution inference, ie statistically grounded of true positive in given clusters, in :func:`nistats.thresholding.map_threshold`
 
 Changes
 -------
@@ -29,7 +30,7 @@ Fixes
 * Explicit method for fixed effects to support image-based summary
   statistics approach.
 * FIR delays are now integers.
-  
+
 Contributors
 ------------